Font sizes are now floats (remove a lot of internal rounding)
And the font Measure is much simplified as it just uses current Size.

The goal is to eventually remove Scale property from TCastleFont
(as it cannot be realized in all descendants, in the way it
currently works; this Scale was supposed to be a helper way to adjust
TTextureFont size, not something adjusted by user).

+++ src/ui/opengl/castlenotifications.pas
@@ -48,7 +48,7 @@ type
       TNotification = class
         Text: string;
         Time: TTimerResult; {< appear time }
-        Width: Integer;
+        Width: Single;
         Color: TCastleColor;
       end;
       TNotificationList = class({$ifdef CASTLE_OBJFPC}specialize{$endif} TObjectList<TNotification>)
